Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
mysql-connector-python==8.2.0
bcrypt==4.0.1
PyPDF2==3.0.1
python-docx==0.8.11
chardet==5.2.0
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
python-dotenv